Move the rest of the operations into the HAL trait.
I don't particularly like how `Platform` permeates everyhing now, but
that'll be a fight for a different day.

And some of the base structures in the crate hierarchy indeed need a
cleanup, as the SEV crates are now leaking into TDX. But that will be a
fight for a different day as well.

Bug: 350496083
Change-Id: I85bd2ce572a2ef784093cc30053aa13a82b42fd3