Implement TDG.VP.VMCALL<GetQuote>
The caller of get_quote should create a 4KiB (or larger) page
before calling get_quote. The input frame must be 4KiB aligned.
The get_quote utility function initializes the request and sends
the shared GPA and size to the hypervisor. Retry at most 3 times.
Response should be processed in a registered handler.

Bug: 358452063
Change-Id: I0457703193f4c82889de639cf6b873da37336443