Bazelify the restricted kernel measurement step
Not sure if we even have a bug for this, but hey, it's done now!

I can't get rid of the `bzimage_provenance_subjects` altogether as
that's also used for the Oak Containers linux kernel, and that's not
built with bazel. Otherwise we could use the same rule to generate the
measurement artifacts for that as well.

Change-Id: Ic43b9b8e96138f0e36eb7f773edf5a5dc26cd1ef