Don't copy the Restricted Kernel procenances to `bin`.
It looks like that in order to access the files from outside the nix
sandbox, we can't rely on `bazel-bin` being availabe; hence, follow the
android client example and copy the binaries into `generated`.

Ideally we'd elide that step as well and just execute one `bazel build`
command...

Also, for the integration test, I had to copy the kernel itself to the
old location. I'm looking at bazelifying that, which would get rid of
that copy as well.

Change-Id: Ic00c02154bb689c98208e9cfd81b75338b585b7c