Print a warning in stage0 if the platform doesn't support GHCB v2
We rely on GHCB MSR protocol in AP bootstrap sequence, but according to
the specs support for that was only added in GHCB protocol version 2.

While GHCB v2 is required for SEV-SNP guests, it is optional for SEV-ES
guests, meaning the AP bootstrap will currently be broken for SEV-ES guests using v1. This will add an explicit note about that.

[ kevinloughlin: massage commit message ]

Bug: 352410101
Change-Id: I69baeb43e92e90ad9892e4bf657662d37af0f91a