impl tdx memory acceptance counter and bug fix
this CR implements a counter similar to SEV's pvalidate status
counter. the counter for tdx includes the number of size mismatch,
pages accepted in 4K or 2M mode. in prod and open source testbed,
I noticed that all pages are accepted in 4K pages.

also it moves some initialization code to the correct location.

Change-Id: I4f0b05411884d7904549d22e555ffb2ba7253f54