Bazelify the Oak Functions integration test
Unfortunately the test is broken right now because the Bazel-built
stage0 doesn't work. But this should line up all the targets and paths
properly so that once stage0 is fixed, the hope is that this will just
magically start to work.

This CR fixes one actual issue in the launcher code: right now if qemu
fails to start for whatever reason (like... broken stage0) write of the
application will stall forever. I've put a 30-second timeout in there,
so that the test will eventually fail if qemu dies.

Change-Id: I064d118df47a0d2114fef6dad650e436609cabf6