Require some references in stage0 to be `'static`
Minor safety change: we store the pointers of these data structures in
the zero page, but there's no guarantee those data structures will stick
around.

From Rust's point of view, this is safe; after all, _taking_ an address
is always safe, it's _dereferencing_ a pointer that's unsafe.

However, for correctness, I think we do want to enforce that the
lifetimes of those data structures are `'static`. Or, well, it'd be even
better if we'd take a `Box` and explicitly handle ownership, but that'd
be a bigger change and I'm not sure it'd be worth the invasive changes.

Change-Id: Id898845867e6afc4611ea61c716c6091d98637e9