Remove dependency vars from BUILD file
This is generally considered an anti-pattern, since it doesn't work well
with some tooling.  For example, it makes updating dependencies using
buildozer comamnds stop working.

While it's not strictly prohibited, it's recommended to avoid it unless
it's truly a pain (many many dependencies, many targets, spread across
files, etc).

For data dependencies, this is easy, we can just create a filegroup.
Unfortunately, for the dependencies, the preferred approach is simply to
repeat the deps.

For more discussion on the matter, see go/build-no-deps-vars

Change-Id: I6a9e9498ea372aeae00e7c78b7f1908076d325c5