Update type_url for event log generated by container orchestrator to ContainerLayerData.
It seems the proto type matches ContainerLayerData and not OrchestratorMeasurements.
We are updating our Oak dependency to the the latest commit, and it seems that
evidence generated by Oak container latest binary is considered as Cb evidence.

The combiniation type_urls generated by stage0, stage1 and orchestrator is
considered as videnceType::CB [1]

[1] https://github.com/project-oak/oak/blob/f5729f4ea324774c4300dc445a38bcf13ce4f400/oak_attestation_verification/src/verifier.rs#L1426

Bug: b/344012112
Change-Id: I494fcf94fb5e38724fda0df416200f49a9432f91