downgrade the system image to stable-20240612
the current debian base image is stable-20240904. it is shipped
with libsystemd0 version `252.22-1~deb12u2`. however, it's
snapshot apt source only provides systemd version `252.22-1~deb12u1`
and we cannot build corresponding system images. apparently it's
a bug of the snapshot service.

stable-20240612 is the last version with systemd version `deb12u1`
and i can confirm the build-base.sh works with it.

Change-Id: I94c812d1008a6d6e5b27f2d7abf83d33b31c3cc0