Build stage0_bin_tdx with Bazel
This is attempt #2 after the previous one was rolled back. This change now moves .stack and .ebda with .data to avoid the data section being placed after .ebda.

This also changes the size of the firmware binary to 2MiB since some versions of QEMU seem to see that as a minimum size.

BUG: 365510289

Change-Id: I2811e29f8922bedb3b670a7f5e83192f16405d38