Implement endorsement verification.
This is following the existing scheme, but is extended to key sets. The existing functions had `_ecdsa` appended to their names so it is clear that they relate to a certain key type.

Even though there are three predicate versions so far, the changes between the predicate versions were simple enough so predicates can be parsed by a single struct hierarchy independent of the type.

Testing is coming up. As of this moment we don't have any producers for the new predicate type.

ACKNOWLEDGE_FAILING_COPYBARA_IMPORT=Seems to have a bad state, will observe

Bug: 366215880
Change-Id: Ie8c9118b89ed62794aa8e29cdf782bbdf418f384