Refactor Stage0 HAL attestation trait
This makes the attestation more generic as a step towards supporting
Intel TDX RTMR-based attestation.

More specifically: for DICE attestation we need to get the attestation
report before we start calling `extend` for the first time. For Intel
TDX and RTMRs we need to get the attestation report when calling
`quote`.

BUG: 368060274

Change-Id: Id7e5270f108bf0ccc556fb0aa9b862ed1096a23e