Crash stage0 if SEV-SNP encrypted bit is suspiciously small.
Right now we just assume it's going to be >32, but let's make that check
explicit and crash if it is not.

This change is tested to the extent that we don't have any platforms
that claim to have the C-bit at < 32, so the crash branch hasn't been
tested, but at least all the existing tests pass.

Bug: 278063057
Change-Id: I64947b16ef8ef9a071efd49df1939330de647ef4