Switch to cargo-audit for dependency checking.
This appears to have no dependencies on cargo, it only requires a lock
file to be present. (If the lock file is absent, it will try to run
cargo).

To test, I nulled out my path; `cargo` was no longer found. Running
`cargo-audit` (by absolute path to its nix location) still returned
results.

This tool does not do license checks for us, so we'll need a different solution for that.

Bug: b/348023135

Change-Id: I5b9732dd0353d101213b375c8c9868db336fe379