Include event log in restricted kernel evidence
Until now the event log was not included in the restricted kernel evidence when converting from a struct to a proto. To make this change non-breaking this adds a new function for doing it and deprecates the old function.

The old function must be removed when all clients have been updated to use the new function.

Change-Id: I746a9ece09cb9b2814f8c055cf9b42f4c82410f1