Add empty config digest to the attestation event.
This fixes an inconsistency between the resulting extracted evidence when using the event logs vs properties in CBOR certificates. The CBOR version puts a 0 length array as the config digest, where the event log completely skipped the config field. This caused errors when verifying the attestation evidence for some cases.

BUG: 372591443

Change-Id: I9a5aadb1a44f25480f350d15f72345a905634e69