Noise implementation for Oak Hello World trusted app
Add an example of the hello world application that uses noise
encryption.

Since we don't yet have any higher-level SDKs, this uses the oak_session
structures directly. This gives us a good reference example in the
repository for using noise, and gives us some context on how we might be
able to make it easier to use for common cases.

I did make an attempt to split out some useful abstractions that we can
eventually consider moving into the SDK, if they really make sense.

Bug: b/343178109

Change-Id: I5ffd58df53ffc075981806100f31157b0ea199df