stage0: Only set the C-bit in non-leaf page table entries
Per AMD docs, page walks are always done as if the C-bit is set,
therefore it doesn't matter whether we set the bit or not.
For Intel TDX, not setting the bit means encrypted (as the C-bit
semantics are inverted), so this doesn't change anything about TDX
behaviour except making it more explicit.

This also mirrors how EDK2 sets up their page tables; they don't bother
setting the C-bit in non-leaf entries either.

Bug: 278063057
Change-Id: Ia32e30aa3bb6d18818796ee4a889d25e6c38bcf9