Add `WBVIND` to stage0 HAL
TDX has special mechanisms for executing that instruction, so it needs
to go into the HAL.

I've left the function as safe because flushing the caches is a safe
operation from Rust's point of view: you can pretty much do it anywhere
with no ill effects on the memory model. Your application will be really
slow, though, but speed isn't a safety issue.

This CR doesn't use the instruction yet; it's just a HAL change.

Bug: 278063057
Change-Id: Ic71aac2c4099bf4f05ae62dcc0eeea63525d57a0