Create a type-safe(r) wrapper around page tables in stage0
The `PageTable` struct from `x86_64` doesn't do any semantic checking of
how you set up your page table beyond the fact that the physical address
is valid.

This CR creates a specific `set_nested_table` method for setting up page
table hierarchies. It's still not _entirely_ type-safe -- we need to
ensure lifetimes make sense and wrap things in `Pin` -- but it's a step
in the direction of making page table ops safer.

Note that the TDX crate does a lot of ops on the `PageTable` structs
from the `x86_64` crate directly; that'll need to be cleaned up at some
point as well.

Bug: 377899703
Change-Id: Ibdcaa1ad096b99aafcb4452c0366e9aa3a54aba6