Add endorsement-based tests as a unit test to module `verifier`.
Since we need to depend on test_util it cannot be an integration test.

For the digests found in the evidence we create and sign endorsements dynamically, and verify by the public keys. Log entries cannot be generated on the fly, but we have one sample of a valid log entry in testdata which happens to match an evidence. So we use that to verify an endorsement which includes a log entry at least once.

Kernel and firmware are not included in this round because the endorsement creation is more complicated. It could be done later as follow-up, but I don't plan on doing this.

This happens now because better test coverage is needed before we can dare to merge go/oak-cr/13683.

Bug: 342041993
Bug: 373849864
Change-Id: I1e1edcd1bf9a9b6776d497b7ef623fe67d75d0a4