Add a simple echo enclave app test
This provides a simple starting point for creating simple tests around
launcher-related functionality.

As part of this, we add the `kill_on_drop` parameter to the launcher, to
allow scope-based management of launched instances. This is particularly
useful in tokio tests, which will hang forever if child tasks are still
spawned when the test exits.

Change-Id: Id0489527b7ead013d5663d0fcfcd4005a75b7c8a