Update initial data load to accept endorsements
The endoresments are not stored anywhere yet. This CR only introduces
the updates to the RK loading protocol that can now accept a proto
containing the endorsements and binary (as well as other things in the
future, if needed).

The Rust launcher code is also updated to provide the application using
the new format (it doesn't provide any endorsements yet).

Discussed in go/on-providing-endorsements

Change-Id: Ic16a346a3d6a6226fff6a1df3515ebfd95db6605