Create C-style binding functions for oak_session types
This CR creates the Rust FFI function interface to
oak_session::ServerSession and oak_session::ClientSession.

It's a fairly close 1:1 mapping to the existing trait methods for the
sessions defined by ProtocolEngine and Session traits.

Error handling is represented using a simple error type that just
contains the message in cases where there was an error; there's no
support yet for error codes or typing.

The tests in this CR are Rust-only, the C-side of the code and tests
will be implemented in a follow-up CR.

Change-Id: I76309858876bf55b2d44ff7e07e50b8bcbc7dcc9