Add an ASAN config for bazel
This will be helpful for making sure we release memory properly when
working with Rust FFI bindings.

This is not yet enabled for CI yet, as there are a few other issues to
work through to make that happen.

(Also removing references to the noci tag in .bazelrc, which is no
longer used by our project).

Fixed: b/382074160

Change-Id: Ida9b2a86b2242c0486aab0ef862360d144640a24