Create a separate `BumpAllocator` for the EBDA space
This replaces the handwritten memory management with existing Rust
`Allocator` trait and our existing `BumpAllocator` implementation.

This reduces the amount of unsafe code; we still do way too much raw
pointer magic for me to be comfortable, and I really don't like how the
`validate()` calls implicity assume the use of `EBDA_ALLOCATOR`, but one
step at a time.

I've also stopped treating the RSDP in a special way. Yes, this may mean
that the RSDP no longer lands exactly in the first kilobyte of EBDA, but
Linux reads the address of the RSDP from the zero page anyway.

Bug: 380246359
Bug: 380857409
Change-Id: Iacd03edf58b430c8b9c48843a506792564ec74f5