Use `PlaintextMessage` for Read/Write methods
My original implementation skipped the wrapper `PlaintextMessage` and
passed the plaintext bytes themselves across the FFI boundary to avoid
unneeded deserialization. But we may need to add additional metadata to
the PlaintextMessage in the future, so to keep things consistent and
avoid API changes later, I'm changing the C/C++ bindings to work with
PlaintextMessage.

Bug: b/337827427
Change-Id: I85587ad738bbaeb739330d41ee7b12634f04ee32