Update SEV attestation report to spec version 1.56
We've got some new fields in the data structure, and that caused the
validation code to fail.

Note that there are fields marked as "Reserved. Must be zero." and
fields marked just as "Reserved."

I've ripped out the code that checks for zero-ness for the "Reserved."
fields, as presumably they _may_ contain data other than zeroes per
spec.

I've also added an explicit check for the attestation report version; if
it changes again, we should get a better error message about it.

...and while I'm in there, it looks like we accidentally treated a u32
field as u64. Oops.

Bug: 384502029

Change-Id: I582140841f1837d466c4f8ace7aed0207899806f