Remove an unnecessary copy of data in `aes_gcm_256_decrypt`
When truncating the vector, the current implementation creates
a copy where we could just truncate in place and not incur the
allocation and copy overhead.

I also tried to get rid of the `Vec::from` in the decryption step but
that just moved the copy into the underlying library so we can't avoid
taht cost. I made a note for future reference about that.

Change-Id: I5b51194fc6b73e30f25ebc7fd18973c4cc9b8d51