Make sure the pointer arrays in Rsdt is aligned correctly.
Previously `Rsdt` is packed so it has alignment of 1. With the `DescriptionHeader`
being 36 bytes, there is no guarantee that the pointer arrays after the
header is 4-byte aligned.

The first attempt is to make `Rsdt` `align(4)`, but this doesn't prevent
the problem from happening again: If an extra byte is added to
`DescriptionHeader`, then `addr_of(rsdt) + size_of(descriptionheader)`
can still return a non-4-byte-aligned address.

By adding a `entries_arr: [u32; 0]`, to `Rsdt`, the compiler will throw an error `entries_arr`
is not 4-byte aligned when the size of `DescriptionHeader` is not 4x bytes.

BUG: 390384333
Change-Id: I1a61250aa0d61b89d2c8d6ff29fae4ea03448701