Extract public keys from the containers event
The keys are extracted by a corresponding policy into the
`EventAttestationResults` proto, and then keys from all events are
placed in a map inside the `AttestationResults` proto.
Each key in a map is marked with a name, which doesn't need to be
globally unique, it just need to be unique within the same evidence
(i.e. for all events).

Bug: 384476430
Bug: 366419879
Change-Id: I4bd9113c7ebd959e56801d50e2952c4c0dd06af4