Ensure that the orchestrator drops the container bundle after unpacking
Once we unpack the archive, there is no need to keep the buffer around
in memory; this can waste a sginificant chunk of memory (we've now got
containers that are 10+ GB in size, that's not a trivial amount).

The logic is that first we convert the `Vec<u8>` into a `bytes::Bytes`
in `lib.rs`, and pass the ownership of that buffer to `run`. Second, in
`run`, the buffer is converted into a `Reader`, and ownership passes to
`archive`; thus, after unpacking the archive, we should ensure it goes
out of scope sooner rather than later.

Bug: 396664122
Change-Id: I791499d35684d87bf10bc38b3ff5111cf511c0d2