Create a data structure for system image that doesn't require copying
In our current implementation as we load the (compressed) data, we just
keep extending a `Vec`. This means that we'll do _a lot_ of
reallocations and data copying as `Vec` needs to be backed by a
contiguous slice of memory.

That memory doesn't really need to be contiguous. This change will
hopefully elide all the reallocations and copying (or, well, most of it)
by just passing around ownership of the chunks of data we load, and the
`Buf` abstraction can happily work on top of non-contiguous slices.

Similar change will need to be made to the orchestrator (for chonky
container images), but that'll be in a separate CR.

Bug: 396664122
Change-Id: Iaeb489ebb11829b6ce37c03d92d99438ef8f0328