Move `Buffer` implementation to `oak_containers/channel`
The `channel` crate is used both by stage1 and the orchestrator, and we
want to use the Buffer in both and it's somewhat connected to
communications so maybe the `channel` crate is good enough a location
for it.

If you think there's a better location for it, I'm open to other ideas.

Bug: 396664122
Change-Id: I3c0a45df2009e4d1a2d346e2edd2b82d774db674