Use `let_chains` to simplify some code in the oak containers launcher
Compared to previous code I think this is way more straightforward, but
the cost here is that we need to enable `let_chains` for this to work.
There is precedent in our code base for using `let_chains`, so I'm not
breaking any new ground here, but feel free to push back if you think
this change is not an improvement over what we already have.

Change-Id: I50d75634e6523e36dd15f67e58b810a9a2c5bf38