Modify oak_session API to accept ownership
This allows us to avoid clone in many cases where it's not really
needed.

In cases where ownership can't be passed, the types can still be cloned
at the call site. So this doesn't prevent users from performing the
clone when needed; it's just explicit now.

Bug: b/394018656
ACKNOWLEDGE_FAILING_COPYBARA_IMPORT=Fixup in cl/729519973

Change-Id: I5cd07aaaa98b70ac45d4662dfbd09cd1a1977174