Expose orchestrator_crypto_client::Signer and make futures Send
This allows callers to use the async InstanceSigner interface and to
provide fakes in tests. Non-Send futures are more difficult to work with
(e.g. Futures awaited from tonic RPC handlers must be Send); since the
only implementation of the interface *does* provide Futures that are
Send, it's easy to add this constraint.

Change-Id: Ib071a6f89ee2f1938288627935302bf6b01b2dde