Add list mode to verify_endorsement script. Change the existing verify mode to work by endorsement hash.
The list mode parses the endorsements and prints information of interest: subject name, subject hash (which is the same as the binary hash), and the number of remaining days the endorsement is valid. So the caller gets a bit of support to select the desired endorsement.

Verify mode now requires the hash of an endorsement (and not the hash of the binary, as before). It downloads the matching verifying key from discoverability storage. Removed the hard-wired dependency on the oak development key.

Bug: 403524037
Change-Id: If4643a65113f3dd9f36de35e07a0504739dfc977