Enable CONFIG_IO_STRICT_DEVMEM in the kernel
This restricts the memory that can be accessed using the /dev/mem
device to PCI config space and the BIOS memory regions. On x86 this
includes the first 1MiB of physical RAM so it still works for our case:
we only use /dev/mem to read the DICE data, which always falls in this
region.

Change-Id: Id2c45eb06a1bc2e28019118a5f161d1a0c75f809