Add basic CxxString support
This is a very simple wrapper around a std::string that makes it
possible to pass C++-allocated string data back to Rust such that Rust
obtains logical ownership of the data.

The test also verifies that BytesView can be used to pass Rust slices
over to C++ code.

A string_view coercion operator is added to BytesView for convenience.

I ran the tests with config=asan as well, since this doesn't happen as
part of our current CI.

Change-Id: Ife842aff6dcfce241702e87f6c6c135324005b20