Separate raw documents from the database.
This CR implements a layered-approach for the database: the first
layer is a meta database that stores all the information except the raw
document content. Instead, the meta database stores ids of the raw
documents. The second layer is a simple key value storage, which stores
the raw documents and the meta database as a encrypted data blob.

Assuming a pair memory id and a blob id is about 20 bytes, this allows users store about 500000 documents, each below 10M.

BUG: 412698203
Change-Id: Ia40fd696547d3279bc13ed58f4374682ec3a6a41