Enable virtio-fs in oak containers kernel
We are testing using virtio-fs to load large amounts of data into our CVMs. We've tested using virtio-fs (rather than virtio-vsock) to load the system image and container bundle and that works with speed ups in loading time screen/45nv8MPnfge9hWd.

We are working on adding dm-verity support to measure the contents of the shared directory at guest boot

Bug: b/415120830
Change-Id: I67142cdef25c78ceda3c1b74edbd4dfaf34238e2