Fix TDX attester to work with Stage 0 logic
I decided here to go in a different direction: rather than creating
unused structs and defaults, I changed to logic to not insist on the
existence of those structs.

I also added a unit test to check that the logic in the main Stage 0 lib
works without panicking.

The additional build targets are needed to disable the global allocator
via a feature, otherwise the test fails when trying to allocate memory.

BUG: 373901407

Change-Id: Ibcf3643352226ee7f08a212ecb516110b3dbd1bf