Drop the half-baked native SDK for Oak Functions
This functionality was speculatively implemented for high-performance
use cases where we would drop the wasm sandbox completely and just call
into native code.

This never ended up used, is not exercised properly (not even in tests!)
and in general goes against the philosophy of Oak Functions, so let's
clean up our code a bit by dropping this peculiar offshoot.

I've left the application configuration protos etc intact, because in
the future we may want to introduce alternative sandboxing methods, and
the plumbing could prove useful at that point. (Also, this means we
don't need to change call sites.)

(Step 1 of 2; deleting the actual cc code will be in a separate PR to
placate Copybara.)

Change-Id: Ifdc8900e117210452ddf6c2208ca009825a111e2