Change unattested sessions to also produce an initial (empty) attest request/response.
This makes an unattested client compatible with self-unidirectional servers. However, it does not make a self-attested client compatible with unattested servers since the client will send a handshake followup with bindings that the server does not expect.

This change will BREAK compatiblity of two unattested peers if they differ on the version of session they use. We don't think this should be a common case, or a case that should occur in production systems as unattested is mostly used for testing or prototyping.

Change-Id: I6e28c0b544c662b81dd8afc3dbd27eb3fcad0ebc