Reserve a separate 1 MiB area for ACPI tables in "high" memory
This change is a bit hacky as I tried to minimize changes to existing
code (e.g. by introducing an allocator). This will require clean-ups in
the future.

The basic problem with EBDA is that it's 128K in size. This works for
QEMU MicroVM, but on Q35 the tables themselves are 128K, and we need to
fit the RSDP in there as well.

I've no idea how the code used to work at all, we must have been
overwriting some memory just past the EBDA.

This code now reserves a chunk of memory just below the 1 GiB mark (or
as close to it as possible, if we've got less memory) and stores the
actual ACPI tables in there. The RSDP stays in the original EBDA.

This is similar to what SeaBIOS is doing.

Note that this doesn't fix the ACPI issues in Q35, but compared to
"baseline" I don't think it breaks it any further either.

Bug: 409562112
Change-Id: Iedcb5886fb7ff6ccea6f14ff4e6c4f1cb80482b5