Create a struct for properly tracking table loader files
This is not in use yet, but will be in a CR or two.

The current code we have treats the ACPI memory rather loosely: we just
happily throw arbitrary buffers of memory around and don't really keep
track of allocations. This also made the TDX code (which needs to
rewrite one ACPI table) more complicated than before, as the TDX code
doesn't care about files, it cares about memory.

My plan in the next CRs is to (a) introduce allocators for the EBDA
memory and the high memory, (b) refactor the table loader code to
properly track the files it's loaded via this new struct, and (c) change
the TDX code to not care about the files at all, but allocate memory for
its new table from the high memory allocator.

Bug: 409562112
Change-Id: I70ab37008c6d191c4a4ade50a911531e9854e906