Track (ACPI) memory and table_loader files separately, properly.
This finally cleans up a lot of the mess around our ACPI code. Benefits
include us now supporting arbitrary files, fewer (mutable) statics and
pointer magic in general.

For the memory, we set up two allocators: one `BumpAllocator` to track
the old EBDA space, and a `LinkedListAllocator` to track the high memory
where the tables themselves go. The MADT rewriter, for example, only
needs to interact with the high memory allocator to allocate space for
the new table, and doesn't need to know about anything else.

The only downside is that we're losing some of the paranoid code in
`acpi_tables.rs` because we can no longer assume everything is in EBDA.
But it's still (much) cleaner.

Bug: 409562112
Change-Id: I3bc053f6a4fee2fd685a146fd70877eb0c11166e