Mount cache/store as volumes when running docke
This makes it easier to iterate on changes to docker dev configs,
without having to refetch-the-world every time you restart the image.

This should be OK for CI runs as well, since both of these cache
directories are themselves inside docker containers. And even if they
did stick around, that wouldn't break hermeticity/reproducibility of the
builds.

Also, consistently use `--mount` syntax, as suggested by the docker
docs. The --volume syntax can lead to unexpected behavior, where you
expect a bind mount, but the source directory doesn't exist, so it falls
back to a volume mount.

Change-Id: Ie8d0f5625d4a81e46039105183f1b2779f0d5e00