Improvement to bazel CI config setting
Setting configs in the justfile is limiting: you have to remember to
apply the BAZEL_CONFIG_FLAG everywhere, and if we happen to run bazel in
other contexts (like other workspaces that we're testing) then the flag
may not be applied.

This CR switch to a new strategy: import a CI-specific bazelrc if it's
present, and then having the CI scripts emit the proper config line
depending on the settings.

This way, things like logging format will always be set correctly, no
matter the context, and you don't have to remember to set --config flags
properly in every CI context you have.

This CR also fixes the double-line log output issue by removing the
tty-related flags in CI mode.

Change-Id: Ied75b43fe3bc31d0f75369facb8b3f197ceea651