Build Linux Kernel as Bazel Nixpkg instead of on flake load
* Prevent the need to unconditionally recompile the Kernel whenver the
  flake is loading. Now the kernel can compile in paralle, when needed
  as part of bazel builds.
* Organize kernel-related things in one place.
* Make the nix kernel compilation build files a little easier to digest.
* Create a simple write_source_files setup for updating Kernel sha256
  sums when the kernel changes. The write_source_files generated test
  serves as our sha consistency check as well.

There's no longer a need for environment variables, and we track the
linux version as a simple text file now (so it can be loaded into
various contexts that need it).

build-base was the consume of the env variables that the flake crated,
so those scripts have been updated as well to use the new approach. I
couldn't find a way to get the path to the already-downloaded linux
sources that nix used, so I use changed the script to download them
again, which only takes a few seconds.

Change-Id: I8bbe22c73f8d80faed268240c0eb80c3666f3619