Add comprehensive doc comments to `oak_session`
This makes the Oak Session SDK more accessible, and easier to use.
All the important structs and traits are documented, and a lot of
context is added, so it becomes clearer what the different components
do and how they fit together.

Some reordering has been made to better align with the structure of the
code.

Change-Id: I927d604268a4e036719c5fc2035864f45e69e044