Be more explicit about artifact release build
Right now, we happen to include release config details in the config=ci
setup.

This makes it easy to miss in the command itself that we really intend
to build a release binary.

It also makes it challenging to run tests in debug mode.

Now that we have relatively few places that we call bazel, it's easier
to include the desired configs in the calls themselves.

Change-Id: I61f30945ee419ae53573448860fb8a2ddc72d561