Return endorsement claims as part of EndorsementDetails.
Other endorsement fields are not desperately needed at this moment: The subject name will be obfuscated anyway, so it doesn't provide any additional information. Similarly, the issued_on field coincides with not_before in virtually all occasions.

Adding the claims in the output increases the size of the result that is passed back to the caller. Need to make sure that this is not on the critical path for Brella.

Bug: 422021058
Change-Id: Ie6884c2f5c6733a1e1e9180a0bab229dfd47f2a5