Fix padding in RomfileCommand union subtypes
This was a tricky one. I recently started noticing stage0_test failing,
but only in `-c dbg` mode, and only when I run locally (not in CI).

After lots of digging into bytes, I found that the three bytes at the
end of the expected struct were changing.

It turns out, the defined struct fields add up to only 125 bytes,
leaving 3 for the compiler to pad however it wants. Thus far, it's
always been 0's, but for some reason, now, on my machine, with `-c dbg`,
the bytes turn into random garbage.

The fix is to update the padding so that we have those bytes covered and
there's not any compiler-controlled padding left.

Each of the sub-types adds up to 124 bytes now (so the total size of
RomfileCommand is 128 byte).

Change-Id: I172bf5c06ec311c1890201dfbd808737f971ccf5