Deprecate irrelevant `tink_endorsement` and `type` fields in the SessionBindingPublicKeyEndorsement
The `tink_endorsement` field was the keystore signature of the binding public key from the
server. This should no longer be used as an endorsement in favor of
relying on the binding key certificate which includes a validity window
for the provided (and signed) key. The entire `oneof type` field should
be removed as all fields in it are no longer referred to and the field
itself does not easily support client migrations from one endorsement
type to another.

BUG: 417151897

Change-Id: I0ec369fa91d1ff7250383146dfe1c1998f9b134a