Verify Rekor log entry by default.
The staging endorsements for Oak are published but don't come with a log entry, so the present change will break that case. On the other hand, all the endorsements in focus right now do have a log entry, so we can make that the default. But the rekor verification still must become a configurable option.

Bug: 423523426
Change-Id: I699c2752948732524fa66dcbfda10fdb151a91b8