Introduce a library for interacting with the Confidential Space TEE and request attestation tokens.
This library includes functionality to:

* Read the attestation token provided by the container launcher
* Request an attestation token from the Confidential Space TEE, including support for nonces.

This implementation currently has some sharp edges. The HTTP client is
hand-coded due to lack of Unix Domain Socket in most HTTP client
libraries for Rust, and is rather brittle. We should look at
implementing it in a more robust way.

The code is poorly (i.e. not) tested since it'd require implementing
also an HTTP server over Unix Domain Sockets.

Bug: 424409724
Change-Id: Ic54b0bd97b16e3ef0758604167d1a0cd1a412a0f