Add `oak_functions_session` service for the standalone app.
We introduce the new `OakFunctionsSession` service. This
includes a single OakSession method which provides the Noise gRPC
endpoint for clients wishing to talk to the standalone Oak Functions
application.

The service could either be provided here or be provided under
`proto/session/` folder where all the HPKE / Noise messages live. We opt
for this service in the `proto/oak_functions/` folder since the folder
provides other OakFunction related services.

Bug: 424407998

Change-Id: Ie956915d3d02066e2444fb3bb7e98d6a4a3f6d8e