Fork oak_functions_containers_app for oak_functions_standalone crate.
Since the Oak Functions Standalone service will be running on a
linux-based environment, a lot of the Oak Functions on Oak Containers
crate can be copied over. After this change, we will start progressively
stripping features from this new crate.

Note that this will not be imported into google3 (see linked change).
This can be changed, if so desired.

Bug: 424407998

Change-Id: Ib4ad31b8db3e68cfcccaa82822a12e7aa36ea511