Strip orchestrator and subtask from oak functions standalone.
Note that this change merely moves the old Oak Functions methods into
the struct as public methods. This is temporary and will
be removed. Instead, the OakFunctionsInstance will be used to carry out
all the WASM and LookupData loading as part of the constructor or a
`initialize()` function.

So far, I have been able to validate that I can run the server locally,
and the OakSession gRPC call returns an error as expected.

BUG: 424407998

Change-Id: Id268057ac626818e1481297c0ad3baf8a6141aa5