Convenience traits for using oak_session in Rust
These traits provide a small but simple boost to usability when working
with Oak Session in Rust.

The tests, along with updated examples, show how the traits can be used.

The traits form the foundation for follow-up usability improvements
where we will provide concrete types that make it difficult or
impossible to attempt to invoke a particular function on the session at
the wrong time.

Change-Id: I6003b3c553e343f8a350afeaabfec1ff033587d5