Add Wasm support for Oak Functions Standalone app.
This change parses the EncryptedMessage from a request, pulls out the
ciphertext, and passes it into the Wasm module. The reply from the Wasm
module is then placed into the EncryptedMessage response to the client.

The test demonstrates a working example of a client connecting with the
server, sending a message, and receiving that message sent back to them.
This is from the echo Wasm module example.

Other features of this change: removed tracing logic from standalone app
(including build dependencies), refactored BUILD dependencies.

Bug: 424407998

Change-Id: I1c8403bc15e8d8dbbc34fcc27866d277d3b0141d