Add encryption for OakSession method in Oak Functions standlone binary.
We introduce the ServerSession to the OakSession with no attestation and
noise NN. This can be changed once we are able to generate an
attestation report (JWT) for the standalone app.

Note that in this case the client is expected to terminate the stream,
which it does in the test example. In the test we use mpsc to manage the
transmission and receiving of data.
https://tokio.rs/tokio/tutorial/channels was used as my reference for
determining how to close the channel.

Bug: 424407998

Change-Id: I1b813473fdc8e9e3691deddc6de772e846960e8f